chore(authentication): users administration improvements

+++ webapp/packages/authentication/src/Administration/UsersManagerService.ts
@@ -9,33 +9,47 @@
 import { injectable } from '@dbeaver/core/di';
 import { CachedResource, GraphQLService, AdminUserInfo } from '@dbeaver/core/sdk';
 
+import { AuthInfoService } from '../AuthInfoService';
 import { AuthProviderService } from '../AuthProviderService';
 
 @injectable()
 export class UsersManagerService {
-  readonly users = new CachedResource([], this.refreshAsync.bind(this), data => !!data.length)
+  readonly users = new CachedResource(
+    [],
+    this.refreshAsync.bind(this),
+    (data, _, userId) => (userId ? data.some(user => user.userId === userId) : !!data.length)
+  )
   constructor(
     private graphQLService: GraphQLService,
     private authProviderService: AuthProviderService,
+    private authInfoService: AuthInfoService,
   ) {
   }
 
-  async create(userId: string): Promise<AdminUserInfo> {
+  async create(userId: string, update?: boolean): Promise<AdminUserInfo> {
     const { user } = await this.graphQLService.gql.createUser({ userId });
 
+    if (update) {
     // TODO: maybe better to do refresh
-    this.users.data.push(user as AdminUserInfo);
+      this.users.data.push(user as AdminUserInfo);
+    }
     return user as AdminUserInfo;
   }
 
   async grantRole(userId: string, roleId: string) {
     await this.graphQLService.gql.grantUserRole({ userId, roleId });
+    await this.users.refresh(userId);
   }
 
-  async delete(userId: string) {
+  async delete(userId: string, update?: boolean) {
+    if (this.authInfoService.userInfo?.userId === userId) {
+      throw new Error('You can\'t delete current logged user');
+    }
     await this.graphQLService.gql.deleteUser({ userId });
-    // TODO: maybe better to do refresh
-    this.users.data.splice(this.users.data.findIndex(user => user.userId === userId), 1);
+
+    if (update) {
+      await this.users.refresh(userId);
+    }
   }
 
   async updateCredentials(userId: string, credentials: Record<string, any>) {
@@ -49,10 +63,25 @@ export class UsersManagerService {
     });
   }
 
-  private async refreshAsync(data: AdminUserInfo[]): Promise<AdminUserInfo[]> {
-    const { users } = await this.graphQLService.gql.getUsersList();
+  private async refreshAsync(
+    data: AdminUserInfo[],
+    _: any,
+    update: boolean,
+    userId?: string
+  ): Promise<AdminUserInfo[]> {
+    const { users } = await this.graphQLService.gql.getUsersList({ userId });
 
-    // TODO: temporary before full implementation was provided
-    return users as AdminUserInfo[];
+    if (!userId) {
+      return users as AdminUserInfo[];
+    }
+
+    const index = data.findIndex(user => user.userId === userId);
+    if (index !== -1) {
+      data.splice(index, 1, ...users as AdminUserInfo[]);
+    } else {
+      data.push(...users as AdminUserInfo[]);
+    }
+
+    return data;
   }
 }
